Filling in functions
Honestly the way the professor is asking us to write the code is not
the greatest way to implement this. It removes the ability to make
multiple queues by requiring a queue global data structure. You
wouldn’t need that global data structure if the DelQueue function
returned a pointer to the new head node instead of the item that was
deleted. Also we need to check boundary conditions for an almost empty
queue, specifically on delqueue.